Actin_segmentation/requirements.txt
tensorflow==1.14.0
Keras==2.3.0
numpy==1.16.5
scikit_image==0.15.0
tqdm==4.36.1
numba==0.45.1
matplotlib==3.1.1
albumentations==0.3.3
ruamel.base==1.0.0
skimage==0.0
